requirements.txt
gradio==4.36.1
requests>=2.28.0
pytrends>=4.9.0
pyyaml>=6.0
pydantic==2.5.3
fastapi==0.104.1
